3PA-Detectron/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
backcall==0.2.0
bartpy==0.0.2
beautifulsoup4==4.11.1
bleach==5.0.1
bokeh==3.0.3
Bottleneck==1.3.5
certifi==2022.12.7
cffi==1.15.1
click==8.0.4
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.0.0
fastjsonschema==2.16.1
Flask==2.2.2
fonttools==4.25.0
imbalanced-learn==0.9.1
imblearn==0.0
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.16.0
jupyter_client==7.3.5
jupyter-contrib-core==0.4.0
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.11.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.5.0
jupyterlab-pygments==0.2.2
kiwisolver==1.4.2
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mistune==2.0.4
munkres==1.1.4
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.5.0
nest-asyncio==1.5.5
notebook==6.4.12
numexpr==2.8.3
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.2.0
pip==23.0.1
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
pure-eval==0.2.2
pybase64==1.2.3
pycparser==2.21
Pygments==2.13.0
pygraphviz==1.9
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pyzmq==23.2.1
scikit-learn==1.1.1
scipy==1.7.3
seaborn==0.12.2
Send2Trash==1.8.0
setuptools==63.4.1
sip==6.6.2
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.5.0
statsmodels==0.13.5
terminado==0.15.0
threadpoolctl==2.2.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.2

traitlets==5.3.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
wheel==0.37.1
wincertstore==0.2
xyzservices==2022.9.0
torchmetrics>=0.9.1
pytorch-lightning
numpy>=1.20.3
wilds>=2.0.0
PyYAML>=6.0
tqdm>=4.64.1
xgboost>=1.6.1
attrs>=21.2.0